requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
openai>=1.0.0
python-dotenv
python-docx